Move characters to character.py
